+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Proper scoping of variables defined within an expand modifier
 - Types via Sorbet
 - Bugfix: multiple modifiers on the same row weren't being handled 
+- Fix modifier string quoting
 - Fix broken Yard doc generation
 
 ## v0.1.2